Restructuring code to be consistent with new nextflow repository structure
