+++ Makefile
@@ -1,65 +1,21 @@
-
-config ?= compileClasspath
-
-ifdef module 
-mm = :${module}:
-else 
-mm = 
-endif 
+# Build the plugin
+assemble:
+	./gradlew assemble
 
 clean:
+	rm -rf .nextflow*
+	rm -rf work
+	rm -rf build
 	./gradlew clean
 
-compile:
-	./gradlew :nextflow:exportClasspath compileGroovy
-	@echo "DONE `date`"
-
-
-check:
-	./gradlew check
-
-
-#
-# Show dependencies try `make deps config=runtime`, `make deps config=google`
-#
-deps:
-	./gradlew -q ${mm}dependencies --configuration ${config}
-
-deps-all:
-	./gradlew -q dependencyInsight --configuration ${config} --dependency ${module}
-
-#
-# Refresh SNAPSHOTs dependencies
-#
-refresh:
-	./gradlew --refresh-dependencies 
-
-#
-# Run all tests or selected ones
-#
+# Run plugin unit tests
 test:
-ifndef class
-	./gradlew ${mm}test
-else
-	./gradlew ${mm}test --tests ${class}
-endif
+	./gradlew test
 
-#
-# generate build zips under build/plugins
-# you can install the plugin copying manually these files to $HOME/.nextflow/plugins
-#
-buildPlugins:
-	./gradlew copyPluginZip
+# Install the plugin into local nextflow plugins dir
+install:
+	./gradlew install
 
-#
-# Upload JAR artifacts to Maven Central
-#
-upload:
-	./gradlew upload
-
-
-upload-plugins:
-	./gradlew plugins:upload
-
-publish-index:
-	./gradlew plugins:publishIndex
+# Publish the plugin
+release:
+	./gradlew releasePlugin
