+++ build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id 'io.nextflow.nextflow-plugin' version '1.0.0-beta.10'
+}
+
+version = '0.4.0'
+
+nextflowPlugin {
+    nextflowVersion = '24.10.0'
+
+    provider = 'phac-nml'
+    className = 'phacnml.plugin.iridanext.IridaNextPlugin'
+    extensionPoints = [
+        'phacnml.plugin.iridanext.IridaNextObserverFactory',
+        'phacnml.plugin.iridanext.SamplesheetParser'
+    ]
+
+}
