--- buildSrc/build.gradle
@@ -1,14 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Support convention plugins written in Groovy. Convention plugins are build scripts in 'src/main' that automatically become available as plugins in the main build.
-    id 'groovy-gradle-plugin'
-}
-
-repositories {
-    // Use the plugin portal to apply community plugins in convention plugins.
-    gradlePluginPortal()
-}
-
