--- buildSrc/src/main/groovy/io.nextflow.groovy-common-conventions.gradle
@@ -1,48 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the groovy Plugin to add support for Groovy.
-    id 'groovy'
-}
-
-repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
-}
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(19)
-    }
-}
-
-compileJava {
-    options.release.set(11)
-}
-
-tasks.withType(GroovyCompile) {
-    sourceCompatibility = '11'
-    targetCompatibility = '11'
-}
-
-tasks.withType(Test) {
-    jvmArgs ([
-            '--add-opens=java.base/java.lang=ALL-UNNAMED',
-            '--add-opens=java.base/java.io=ALL-UNNAMED',
-            '--add-opens=java.base/java.nio=ALL-UNNAMED',
-            '--add-opens=java.base/java.nio.file.spi=ALL-UNNAMED',
-            '--add-opens=java.base/java.net=ALL-UNNAMED',
-            '--add-opens=java.base/java.util=ALL-UNNAMED',
-            '--add-opens=java.base/java.util.concurrent.locks=ALL-UNNAMED',
-            '--add-opens=java.base/java.util.concurrent.atomic=ALL-UNNAMED',
-            '--add-opens=java.base/sun.nio.ch=ALL-UNNAMED',
-            '--add-opens=java.base/sun.nio.fs=ALL-UNNAMED',
-            '--add-opens=java.base/sun.net.www.protocol.http=ALL-UNNAMED',
-            '--add-opens=java.base/sun.net.www.protocol.https=ALL-UNNAMED',
-            '--add-opens=java.base/sun.net.www.protocol.ftp=ALL-UNNAMED',
-            '--add-opens=java.base/sun.net.www.protocol.file=ALL-UNNAMED',
-            '--add-opens=java.base/jdk.internal.misc=ALL-UNNAMED',
-    ])
-}
