--- buildSrc/src/main/groovy/io.nextflow.groovy-library-conventions.gradle
@@ -1,11 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the common convention plugin for shared build configuration between library and application projects.
-    id 'io.nextflow.groovy-common-conventions'
-    // Apply the java-library plugin for API and implementation separation.
-    id 'java-library'
-}
-
