--- launch.sh
@@ -1,2 +1,0 @@
-export NXF_PLUGINS_DEV=$PWD/plugins
- ../nextflow/launch.sh "$@"
