--- nextflow.config
@@ -1,3 +1,0 @@
-plugins {
-  id 'nf-iridanext'
-}
