--- plugins/nf-iridanext/build.gradle
@@ -1,89 +1,0 @@
-/*
- * Copyright 2021-2022, Seqera Labs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    // Apply the groovy plugin to add support for Groovy
-    id 'io.nextflow.groovy-library-conventions'
-    id 'idea'
-}
-
-group = 'io.nextflow'
-// DO NOT SET THE VERSION HERE
-// THE VERSION FOR PLUGINS IS DEFINED IN THE `/resources/META-INF/MANIFEST.NF` file
-
-idea {
-    module.inheritOutputDirs = true
-}
-
-repositories {
-    mavenCentral()
-    maven { url = 'https://jitpack.io' }
-    maven { url = 'https://s3-eu-west-1.amazonaws.com/maven.seqera.io/releases' }
-    maven { url = 'https://s3-eu-west-1.amazonaws.com/maven.seqera.io/snapshots' }
-}
-
-configurations {
-    // see https://docs.gradle.org/4.1/userguide/dependency_management.html#sub:exclude_transitive_dependencies
-    runtimeClasspath.exclude group: 'org.slf4j', module: 'slf4j-api'
-}
-
-sourceSets {
-    main.java.srcDirs = []
-    main.groovy.srcDirs = ['src/main']
-    main.resources.srcDirs = ['src/resources']
-    test.groovy.srcDirs = ['src/test']
-    test.java.srcDirs = []
-    test.resources.srcDirs = ['src/testResources']
-}
-
-ext{
-    nextflowVersion = '23.04.0'
-}
-
-dependencies {
-    // This dependency is exported to consumers, that is to say found on their compile classpath.
-    compileOnly "io.nextflow:nextflow:$nextflowVersion"
-    compileOnly 'org.slf4j:slf4j-api:1.7.10'
-    compileOnly 'org.pf4j:pf4j:3.4.1'
-
-    // add here plugins depepencies
-    implementation 'net.jimblackler.jsonschemafriend:core:0.12.3'
-
-    // test configuration
-    testImplementation "org.codehaus.groovy:groovy:3.0.17"
-    testImplementation "org.codehaus.groovy:groovy-nio:3.0.17"
-    testImplementation "io.nextflow:nextflow:$nextflowVersion"
-    testImplementation ("org.codehaus.groovy:groovy-test:3.0.17") { exclude group: 'org.codehaus.groovy' }
-    testImplementation ("cglib:cglib-nodep:3.3.0")
-    testImplementation ("org.objenesis:objenesis:3.1")
-    testImplementation ("org.spockframework:spock-core:2.2-groovy-3.0") { exclude group: 'org.codehaus.groovy'; exclude group: 'net.bytebuddy' }
-    testImplementation ('org.spockframework:spock-junit4:2.2-groovy-3.0') { exclude group: 'org.codehaus.groovy'; exclude group: 'net.bytebuddy' }
-    testImplementation ('com.google.jimfs:jimfs:1.1')
-
-    testImplementation(testFixtures("io.nextflow:nextflow:$nextflowVersion"))
-    testImplementation(testFixtures("io.nextflow:nf-commons:$nextflowVersion"))
-
-    // see https://docs.gradle.org/4.1/userguide/dependency_management.html#sec:module_replacement
-    modules {
-        module("commons-logging:commons-logging") { replacedBy("org.slf4j:jcl-over-slf4j") }
-    }
-}
-
-// use JUnit 5 platform
-test {
-    useJUnitPlatform()
-}
-
