--- plugins/nf-iridanext/src/resources/META-INF/MANIFEST.MF
@@ -1,6 +1,0 @@
-Manifest-Version: 1.0
-Plugin-Id: nf-iridanext
-Plugin-Version: 0.3.0
-Plugin-Class: nextflow.iridanext.IridaNextPlugin
-Plugin-Provider: nextflow
-Plugin-Requires: >=22.10.0
