--- plugins/nf-iridanext/src/resources/META-INF/extensions.idx
@@ -1,2 +1,0 @@
-nextflow.iridanext.IridaNextObserverFactory
-nextflow.iridanext.SamplesheetParser
