+++ src/main/groovy/phacnml/plugin/iridanext/MetadataParserJSON.groovy
@@ -1,0 +1,29 @@
+package phacnml.plugin.iridanext
+
+import java.nio.file.FileSystems
+import java.nio.file.PathMatcher
+import java.nio.file.Path
+import groovy.json.JsonSlurper
+
+import groovy.transform.CompileStatic
+
+import nextflow.Nextflow
+import phacnml.plugin.iridanext.MetadataParser
+
+@CompileStatic
+class MetadataParserJSON extends MetadataParser {
+
+    private JsonSlurper jsonSlurper
+    
+    public MetadataParserJSON(PathMatcher pathMatcher = null) {
+        super(pathMatcher)
+
+        this.jsonSlurper = new JsonSlurper()
+    }
+
+    @Override
+    public Map<String, Object> parseMetadata(Path path) {
+        path = Nextflow.file(path) as Path
+        return jsonSlurper.parseText(path.text) as Map<String, Object>
+    }
+}
