+++ src/main/resources/META-INF/extensions.idx
@@ -1,0 +1,2 @@
+phacnml.plugin.iridanext.IridaNextObserverFactory
+phacnml.plugin.iridanext.SamplesheetParser
