+++ src/main/resources/phacnml/plugin/iridanext/output_schema.json
